pyproject.toml
[tool.poetry]
name = "begonia"
version = "0.1.0"
description = ""
authors = ["vforfreedom <vforfreedom96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
protobuf = "^4.21.12"
grpcio-tools = "^1.50.0"
grpcio = "^1.62.2"
httpx = "^0.27.0"
mypy-protobuf = "^3.6.0"
google-api-python-client = "^2.133.0"
pydantic-core = "^2.19.0"
annotated-types = "^0.7.0"
pydantic = "^2.7.4"
typing-extensions = "^4.12.2"
loguru = "^0.7.2"
h11 = "^0.14.0"
grpcio-status = "1.42.0"


[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
